requirements/base.txt
fastapi==0.92.0
fastapi_utils==0.2.1
numpy==1.24.2
requests==2.27.1
opencv-contrib-python==4.7.0.68
python-multipart==0.0.5
uvicorn==0.20.0
tqdm==4.64.1
pytz==2019.3
pillow==9.3.0
apscheduler==3.9.1.post1
yoloface==0.0.4